changing jparepository to native repository

+++ src/main/java/com/enigma/traveloca/service/impl/CustomerServiceImpl.java
@@ -3,6 +3,7 @@ package com.enigma.traveloca.service.impl;
 import com.enigma.traveloca.dto.request.create.CreateCustomerRequest;
 import com.enigma.traveloca.dto.request.update.UpdateCustomerRequest;
 import com.enigma.traveloca.dto.response.CustomerResponse;
+import com.enigma.traveloca.entity.AppUser;
 import com.enigma.traveloca.entity.Customer;
 import com.enigma.traveloca.entity.UserCredential;
 import com.enigma.traveloca.repository.CustomerRepository;
@@ -11,6 +12,8 @@ import com.enigma.traveloca.service.CustomerService;
 import com.enigma.traveloca.utils.ValidationUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
@@ -36,20 +39,27 @@ public class CustomerServiceImpl implements CustomerService {
     @Override
     public CustomerResponse save(CreateCustomerRequest request){
         validationUtil.validate(request);
-        UserCredential userCredential = userCredentialRepository.findById(request.getUserCredentialId())
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User credential not found"));
+        UserCredential userCredential = userCredentialRepository.findById(request.getUserCredentialId());
+
         Customer customer = Customer.builder()
                 .name(request.getName())
                 .userCredential(userCredential)
                 .build();
-        repository.saveAndFlush(customer);
+        repository.save(customer);
         return mapToResponse(customer);
     }
 
     @Transactional(readOnly = true)
     @Override
     public CustomerResponse findById(String id){
-        Customer customer = findByIdOrThrowException(id);
+        Customer customer = repository.findById(id);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        AppUser appUser = (AppUser) authentication.getPrincipal();
+
+        if (!appUser.getId().equals(customer.getUserCredential().getId()) && !appUser.getRole().name().equals("ROLE_ADMIN")) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "unauthorized");
+        }
 
         return mapToResponse(customer);
     }
@@ -57,13 +67,20 @@ public class CustomerServiceImpl implements CustomerService {
     @Transactional(readOnly = true)
     @Override
     public Customer getById(String id) {
-        return findByIdOrThrowException(id);
+        return repository.findById(id);
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void deleteById(String id) {
-        Customer customer = findByIdOrThrowException(id);
+        Customer customer = repository.findById(id);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        AppUser appUser = (AppUser) authentication.getPrincipal();
+
+        if (!appUser.getId().equals(customer.getUserCredential().getId()) && !appUser.getRole().name().equals("ROLE_ADMIN")) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "unauthorized");
+        }
 
         repository.delete(customer);
     }
@@ -71,7 +88,7 @@ public class CustomerServiceImpl implements CustomerService {
     @Transactional(rollbackFor = Exception.class)
     @Override
     public CustomerResponse update(UpdateCustomerRequest request) {
-        Customer customer = findByIdOrThrowException(request.getId());
+        Customer customer = repository.findById(request.getId());
 
         Customer updated = Customer.builder()
                 .id(customer.getId())
@@ -79,12 +96,9 @@ public class CustomerServiceImpl implements CustomerService {
                 .userCredential(customer.getUserCredential())
                 .build();
 
-        return mapToResponse(repository.saveAndFlush(updated));
+        return mapToResponse(repository.update(updated));
     }
 
-    private Customer findByIdOrThrowException(String id) {
-        return repository.findById(id).orElseThrow(() -> { throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Customer not found");});
-    }
     private CustomerResponse mapToResponse(Customer customer) {
         String userCredentialId;
         if (customer.getUserCredential() == null)  userCredentialId = null;
